+++ routes/web.php
@@ -230,6 +230,7 @@ Route::get('/modificarContrato/{id}', [ContratoController::class, 'edit'])->midd
 Route::get('/testContrato/{id}', [ContratoController::class, 'test'])->middleware('auth');
 Route::get('/modificarContratroNoPaga/{id}', [ContratoController::class, 'updateNoPaga'])->middleware('auth');
 Route::patch('/modificarContratoCliente', [ContratoController::class, 'updateCliente'])->middleware('auth');
+Route::get('/modificarContratoDireccion', [DireccionController::class, 'create'])->middleware('auth');
 Route::patch('/modificarContratoCoordenadas', [ContratoController::class, 'updateCoordenadas'])->middleware('auth');
 Route::patch('/modificarContrato', [ContratoController::class, 'update'])->middleware('auth');
 Route::patch('/realtaContrato', [ContratoController::class, 'undestroy'])->middleware('auth');